STEAM-617: Implement cluster functions in service
Additionally, this commit removes temporary sql package.
