+++ master/data/constants.go
@@ -1,4 +1,21 @@
-package sql
+/*
+	Copyright (C) 2016 H2O.ai, Inc. <http://h2o.ai/>
+
+	This program is free software: you can redistribute it and/or modify
+	it under the terms of the GNU Affero General Public License as
+	published by the Free Software Foundation, either version 3 of the
+	License, or (at your option) any later version.
+
+	This program is distributed in the hope that it will be useful,
+	but WITHOUT ANY WARRANTY; without even the implied warranty of
+	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+	GNU Affero General Public License for more details.
+
+	You should have received a copy of the GNU Affero General Public License
+	along with this program.  If not, see <http://www.gnu.org/licenses/>.
+*/
+
+package data
 
 const (
 	// --- History ---
@@ -18,7 +35,7 @@ const (
 	View = "view"
 
 	// --- Role ---
-	Superuser = "superuser"
+	SuperuserRN = "superuser"
 )
 
 var CLUSTER_TYPES = []string{
@@ -222,6 +239,36 @@ var STATES = []string{
 	"disconnected",
 	"failed",
 	"completed",
+}
+
+type states struct {
+	Idle         string
+	Starting     string
+	Started      string
+	Suspending   string
+	Suspended    string
+	Stopping     string
+	Stopped      string
+	Blocked      string
+	Disconnected string
+	Failed       string
+	Completed    string
+}
+
+func initState() states {
+	return states{
+		"idle",
+		"starting",
+		"started",
+		"suspending",
+		"suspended",
+		"stopping",
+		"stopped",
+		"blocked",
+		"disconnected",
+		"failed",
+		"completed",
+	}
 }
 
 type stateKeys struct {
