+++ master/data/crud.go
@@ -1,6 +1,6 @@
 // GENERATED BY crudgen; DON'T EDIT
 
-package sql
+package data
 
 import (
 	"database/sql"
