--- master/web/model_test.go
@@ -1,94 +1,0 @@
-/*
-  Copyright (C) 2016 H2O.ai, Inc. <http://h2o.ai/>
-
-  This program is free software: you can redistribute it and/or modify
-  it under the terms of the GNU Affero General Public License as
-  published by the Free Software Foundation, either version 3 of the
-  License, or (at your option) any later version.
-
-  This program is distributed in the hope that it will be useful,
-  but WITHOUT ANY WARRANTY; without even the implied warranty of
-  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-  GNU Affero General Public License for more details.
-
-  You should have received a copy of the GNU Affero General Public License
-  along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-package web
-
-import "testing"
-
-// TODO: This function is dependent upon a more robust h2o script running, this
-// will need to have multiple models of all types running in the cluster.
-func TestModelCRUD(tt *testing.T) {
-	t := newTest(tt)
-
-	// -- Setup --
-
-	projectId, err := t.svc.CreateProject(t.su, "project1", "test project", "")
-	t.nil(err)
-	clusterId, err := t.svc.RegisterCluster(t.su, ClusterAddress)
-	t.nil(err)
-
-	// -- C --
-
-	modelId := make([]int64, len(h2oModels))
-	for i, model := range h2oModels {
-		var err error
-		// If no name is supplied, model should inherit name from H2O key
-		modelId[i], err = t.svc.ImportModelFromCluster(t.su, clusterId, projectId, model.name, "")
-		t.nil(err)
-	}
-
-	// -- R --
-
-	// model, err := t.svc.GetModel(t.su, modelId)
-	// t.nil(err)
-
-	// t.ok(name == model.Name, "GetModel: Name: expected %s got %s", name, model.Name)
-	// t.ok(h2oModelKey == model.ModelKey, "GetModel: ModelKey: expected %s got %s", h2oModelKey, model.ModelKey)
-
-	// TODO VERIFY MODEL INFORMATION FROM H2O
-
-	// TODO: Deprecated?
-	models, err := t.svc.GetModels(t.su, projectId, 0, 10000)
-	t.nil(err)
-	t.log(models)
-
-	// t.ok(name == models[0].Name, "GetModel: Name: expected %s got %s", name, models[0].Name)
-	// t.ok(h2oModelKey == models[0].ModelKey, "GetModel: ModelKey: expected %s got %s", h2oModelKey, models[0].ModelKey)
-
-	// TODO MORE TESTING HERE DEPENDENT ON H2O SCRIPT
-	binModels, err := t.svc.FindModelsBinomial(t.su, projectId, "", "", true, 0, 1000)
-	t.nil(err)
-	t.log(binModels)
-
-	// binModel, err := t.svc.GetModelBinomial(t.su, binModelId)
-	// t.nil(err)
-	// t.log(binModel)
-
-	mulModels, err := t.svc.FindModelsMultinomial(t.su, projectId, "", "", true, 0, 1000)
-	t.nil(err)
-	t.log(mulModels)
-
-	// mulModel, err := t.svc.GetModelMultinomial(t.su, mulModelId)
-	// t.nil(err)
-	// t.log(mulModel)
-
-	regModels, err := t.svc.FindModelsRegression(t.su, projectId, "", "", true, 0, 1000)
-	t.nil(err)
-	t.log(regModels)
-
-	// regModel, err := t.svc.GetModelRegression(t.su, regModelId)
-	// t.nil(err)
-	// t.log(regModel)
-
-	// -- U --
-
-	// -- D --
-	for _, id := range modelId {
-		err := t.svc.DeleteModel(t.su, id)
-		t.nil(err)
-	}
-}
